rename impersonate to delegate
