+++ pkg/delegate/repository.go
@@ -1,4 +1,4 @@
-package impersonate
+package delegate
 
 import (
 	"context"
